build and layout cleanup
Create a separation of concerns:

* src (circ bot related code)
* log (log module, used by libirc and circ)
* libirc (irc protocol and message library)
* thirdparty (shit we steal)

Bring cmake code up to snuff.
