--- build.sh
@@ -1,7 +1,0 @@
-#!/bin/sh
-
-mkdir -p build
-cd ./build
-
-cmake ../
-make
