--- clean.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-rm -rf ./build/
