--- lib/log/log.h
@@ -1,5 +1,0 @@
-void
-log_info (char* fmt, ...);
-
-void
-log_debug (char* fmt, ...);
