+++ libirc/hooks.c
