+++ libirc/ircium-parser/ircium-message.h
