+++ log/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(log log.c)
+
+target_include_directories(log PUBLIC ..)
