+++ log/log.c
@@ -1,5 +1,4 @@
 #include "log.h"
-#include "config/config.h"
 #include <stdarg.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -7,14 +6,12 @@
 void
 log_debug (char* fmt, ...)
 {
-    struct ConfigType *config = get_config();
-
-	va_list argptr;
-	va_start (argptr, fmt);
-	if (config->debug) {
+	if (DEBUG) {
+		va_list argptr;
+		va_start (argptr, fmt);
 		vprintf (fmt, argptr);
+		va_end (argptr);
 	}
-	va_end (argptr);
 }
 
 void
